[Diamond IV] Title: 카드 게임 2, Time: 376 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Diamond/28039. 카드 게임 2/카드 게임 2.cc
@@ -48,7 +48,7 @@ void solve()
                 arr.pop_back();
                 arr.pop_back();
                 arr.pop_back();
-                arr.push_back((long long)((__int128)a - b + c));
+                arr.push_back(a - b + c);
             }
             else
                 break;
